Updated copyright year and project url. Added support for AMF3 Aggregate type. Implemented new handshake processes to support RTMPE etc.
git-svn-id: http://red5.googlecode.com/svn/java/server/trunk@4154 1b6495e4-3631-0410-8e05-8f51eee8b9cc

+++ src/org/red5/server/net/rtmp/RTMPMinaConnection.java
@@ -1,9 +1,9 @@
 package org.red5.server.net.rtmp;
 
 /*
- * RED5 Open Source Flash Server - http://www.osflash.org/red5
+ * RED5 Open Source Flash Server - http://code.google.com/p/red5/
  *
- * Copyright (c) 2006-2009 by respective authors (see below). All rights reserved.
+ * Copyright (c) 2006-2010 by respective authors (see below). All rights reserved.
  *
  * This library is free software; you can redistribute it and/or modify it under the
  * terms of the GNU Lesser General Public License as published by the Free Software
@@ -142,11 +142,11 @@ public class RTMPMinaConnection extends RTMPConnection implements RTMPMinaConnec
                 		int downStream = filter.getMaxReadThroughput();
                 		int upStream = filter.getMaxWriteThroughput();
                 		if (downStream > 0) {
-                			ServerBW serverBW = new ServerBW(downStream / 8);
+                			ServerBW serverBW = new ServerBW((int) downStream / 8);
                 			getChannel(2).write(serverBW);
                 		}
                 		if (upStream > 0) {
-                			ClientBW clientBW = new ClientBW(upStream / 8, (byte) 0);
+                			ClientBW clientBW = new ClientBW((int) upStream / 8, (byte) 0);
                 			getChannel(2).write(clientBW);
             				// Update generation of BytesRead messages
             				// TODO: what are the correct values here?
